% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activateAnnotation.R
\name{activateAnnotation}
\alias{activateAnnotation}
\title{activateAnnotation - Get the right annotation (possibly installing it) for an expressionset. It will load the package.}
\usage{
activateAnnotation(eset)
}
\arguments{
\item{eset}{The expression set that has a corresponding (affy) annotation.}
}
\value{
The name of the expression set.
}
\description{
activateAnnotation - Get the right annotation (possibly installing it) for an expressionset. It will load the package.
}
