% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSubset.R
\name{featureSubset}
\alias{featureSubset}
\title{Reduce an ExpressionSet by a given list of ids. Note the ids must match
the featureNames of the ExpressionSet. The result is an ExpressionSet
limited to only those features in "id" list.}
\usage{
featureSubset(eset, ids, debug = TRUE)
}
\arguments{
\item{eset}{An ExpressionSet to subset}

\item{ids}{A vector of featureNames to subset the ExpressionSet on.}
}
\value{
A new ExpressionSet reduced by the features in 'ids'.
}
\description{
Reduce an ExpressionSet by a given list of ids. Note the ids must match
the featureNames of the ExpressionSet. The result is an ExpressionSet
limited to only those features in "id" list.
}
