% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnotation.R
\name{getAnnotation}
\alias{getAnnotation}
\title{getAnnotation - get the annotation table for a given ExpressionSet}
\usage{
getAnnotation(eset)
}
\arguments{
\item{eset}{The expression set that has a corresponding (affy) annotation.}
}
\value{
The name of the annotation map. Store this in saeAffyAnnotation variable to get automatic access for subsequent calls.
}
\description{
getAnnotation - get the annotation table for a given ExpressionSet
}
