% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnotationByEntrezID.R
\name{getAnnotationByEntrezID}
\alias{getAnnotationByEntrezID}
\title{getAnnotationByEntrezID - get the annotation table of EntrezID's for a given ExpressionSet}
\usage{
getAnnotationByEntrezID(eset)
}
\arguments{
\item{eset}{The expression set that has a corresponding (affy) annotation.}
}
\value{
A list of annotations (ENTREZID) by probeset.
}
\description{
getAnnotationByEntrezID - get the annotation table of EntrezID's for a given ExpressionSet
}
