% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnotationByEnsembl.R,
%   R/getAnnotationBySymbol.R
\name{getAnnotationBySymbol}
\alias{getAnnotationBySymbol}
\title{getAnnotationByEnsembl - get the annotation table of ensembl ID's for a given ExpressionSet}
\usage{
getAnnotationBySymbol(eset)

getAnnotationBySymbol(eset)
}
\arguments{
\item{eset}{The expression set that has a corresponding (affy) annotation.}
}
\value{
A list of annotations (ENSEMBL) by probeset.

A list of annotations (SYMBOL) by probeset.
}
\description{
getAnnotationByEnsembl - get the annotation table of ensembl ID's for a given ExpressionSet

getAnnotationBySymbol - get the annotation table of gene symbols for a given ExpressionSet
}
