% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnotationList.R
\name{getAnnotationList}
\alias{getAnnotationList}
\title{getAnnotationList - get the annotation list for a given type}
\usage{
getAnnotationList(eset, mapping)
}
\arguments{
\item{eset}{The expression set to get annotation for.}

\item{mapping}{The specific field to get (e.g. SYMBOL, ENTREZID) as a string.}
}
\value{
A list of annotations by probeset.
}
\description{
getAnnotationList - get the annotation list for a given type
}
