% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExpressionSetByEntrezID.R
\name{getExpressionSetByEntrezID}
\alias{getExpressionSetByEntrezID}
\title{This function takes an expression set (assumed to be Affymetrix), and reduces the features (probesets) to only one
entry per gene id (EntrezID). The featurenames of the returned expression set are the EntrezID's.
Currently, this is only implemented for taking the highest median expression if there
are duplicates.}
\usage{
getExpressionSetByEntrezID(eset)
}
\arguments{
\item{eset}{\itemize{
\item An expression set to reduce to EntrezID reporters.
}}
}
\value{
A modified expression set (based on eset) with only features that are EntrezIDs.
}
\description{
TODO: Add more ways of summarizing.
}
